move tag_burst_from_moving_average_baseline() into tagging.py
